+++ TeamProject/GlobalWarmingGame/Interactions/Interactables/Buildings/BuildingManager.cs
